Fe create admin panel & fix ap is (#26)
* fetch current user

* disable drag option in links

* fixed header responsiveness

* Created new index page

* fixed home page

* DB - Update schema with modified table

Updated the schema.sql file to reflect the modified table definitions for the `users`, `teachers`, and `students` tables.

In the `users` table, the `teacherId` column has been set to auto-increment and is now referenced as a foreign key. Additionally, the `id` column has been removed from the `teachers` and `students` tables, and the `userId` and `teacherId` columns have been renamed to `teacherId` and `studentId` respectively.

These changes ensure that the `teacherId` and `studentId` columns in the `teachers` and `students` tables reference the `id` column of the `users` table appropriately.

Furthermore, triggers have been added to automatically update the `teachers` table when a teacher is added or removed, and to update the `students` table when a student is added or removed. Additionally, a trigger has been implemented to update the `studentCount` in the `teachers` table when a student is added or deleted.

Please note that using the `teacherId` as both an auto-incremented column and a foreign key in the `users` table may require additional consideration to avoid conflicts.

* DB and BE - fix register user

* removed isStudent from BE

* Admin Adjusments

* Fix/fe 40 f ix all api models to {status, data} (#25)

* wip backend and frontend

* fix(FE-40): Updated model interface

* fix(FE): Updated username on admin panel

* fix(FE-40): Fix unexpected profile error

* Updated Readme - [WIP]

* feat-FE- Admin deletes user in dashboards

* feat-FE- Admin registers user in dashboards - [WIP]

* feat-FE- Update table list when admin deletes user

* feat-FE- Hide current admin user

1. Hide current admin user in dashboards
2. If selected user has no id, display error notification
3. Display message if list is empty